# ./block_writer -prepare uses prepared inserts