This should be useful to compare the prepare path to the exec path. It
seems to improve performance on a local one-node test cluster by a small
amount at the moment.

```
benchstat exec.txt prepare.txt
name                 old time/op  new time/op  delta
BlockWriter           212µs ± 0%   196µs ± 0%   ~     (p=1.000 n=1+1)
BlockWriterBatch10    309µs ± 0%   261µs ± 0%   ~     (p=1.000 n=1+1)
BlockWriterBatch100   242µs ± 0%   229µs ± 0%   ~     (p=1.000 n=1+1)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/examples-go/105)
<!-- Reviewable:end -->