@petermattis 2017-04-23
I didn't look at this closely yet, but wanted to point out that `cockroachdb/loadgen/kv` uses prepared statements and other operations and is seeing all of the new development. I've been meaning to get rid of `block_writer`. Perhaps we should add a WARNING comment to the file pointing to `cockroachdb/loadgen/kv`. 

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/examples-go/105#-:-KiNEkHYrE13ZVo8fp6w:b-p4bz3p)*
<!-- Sent from Reviewable.io -->