@jseldess 2017-04-23
@petermattis, if we do get rid of `block_writer`, please let me know a bit ahead of time, as we reference it from one of our "core feature" tutorials: https://www.cockroachlabs.com/docs/demo-automatic-rebalancing.html