@petermattis 2017-04-24
@jseldess Roger. There isn't an imminent plan to get rid of `block_writer`. I'd just prefer to see new developments happen on `loadgen/kv`.