@jordanlewis 2017-05-03
Whoops, yeah, should have just used `kv`. I'll close this. I'll probably make a patch to `kv` that optionally doesn't use prepared statements since I think that's still a useful point of comparison.